+++ eyes.images.java/src/test/java/com/applitools/eyes/images/TestImageUtils.java
@@ -13,16 +13,20 @@ import java.util.ArrayList;
 
 public class TestImageUtils extends ReportingTestSuite {
 
-    public class TestLogHandler implements LogHandler {
+    public static class TestLogHandler extends LogHandler {
 
-        private ArrayList<String> messages = new ArrayList<>();
+        private final ArrayList<String> messages = new ArrayList<>();
+
+        protected TestLogHandler() {
+            super(false);
+        }
 
         @Override
         public void open() {}
 
         @Override
-        public void onMessage(boolean verbose, String logString) {
-            messages.add(logString);
+        public void onMessage(String message) {
+            messages.add(message);
         }
 
         @Override
@@ -61,7 +65,7 @@ public class TestImageUtils extends ReportingTestSuite {
         BufferedImage cropped = ImageUtils.cropImage(logger, image, new Region(600, 350, 300, 300));
         Assert.assertEquals(cropped.getWidth(), 200, "widths differ");
         Assert.assertEquals(cropped.getHeight(), 150, "heights differ");
-        Assert.assertTrue(testLogHandler.contains("[LOG    ] {} [1] com.applitools.utils.ImageUtils.cropImage(): WARNING - requested cropped getArea overflows image boundaries."));
+        Assert.assertTrue(testLogHandler.contains("[Notice]\t{} [1] com.applitools.utils.ImageUtils.cropImage(): WARNING - requested cropped getArea overflows image boundaries."));
     }
 
     @Test
@@ -73,6 +77,6 @@ public class TestImageUtils extends ReportingTestSuite {
         BufferedImage cropped = ImageUtils.cropImage(logger, image, new Region(850, 100, 300, 200));
         Assert.assertEquals(cropped.getWidth(), 800, "widths differ");
         Assert.assertEquals(cropped.getHeight(), 500, "heights differ");
-        Assert.assertTrue(testLogHandler.contains("[LOG    ] {} [1] com.applitools.utils.ImageUtils.cropImage(): WARNING - requested cropped getArea results in zero-size image! Cropped not performed. Returning original image."));
+        Assert.assertTrue(testLogHandler.contains("[Notice]\t{} [1] com.applitools.utils.ImageUtils.cropImage(): WARNING - requested cropped getArea results in zero-size image! Cropped not performed. Returning original image."));
     }
 }