+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 [vNext]
 ### Added
 - A new API for closing batch explicitly: `BatchClose`. [Trello 2189](https://trello.com/c/SlHH9Ssb)
+- A new log handler for sending logs to the eyes server. [Trello 2206](https://trello.com/c/EX8JfK7W)
 
 [3.180.0] - 2020-10-09
 ### Updated